pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Build a RAG End-End Medical Chatbot with LLMs, LangChain, Pinecone, Flask"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==3.1.1",
    "langchain==0.3.26",
    "langchain-community==0.3.26",
    "langchain-google-genai==2.0.7",
    "langchain-openai==0.3.24",
    "langchain-pinecone==0.2.8",
    "pypdf==5.6.1",
    "python-dotenv==1.0.0",
    "sentence-transformers==4.1.0",
]
